refactor(checker): split checker.rs into per-language modules
src/checker.rs held 25 Checker trait impls plus 11 local macros and a
large inline test module — 2038 production SLOC, 2.5x the 800-line
self-scan limit. Split it per-language like src/getter and
src/metrics/abc: checker.rs keeps the Checker trait, the 11 macros, the
3 cross-module-shared pub(crate) helpers (java_anonymous_class_body,
csharp_accessor_count, csharp_member_has_accessors), the Rust
test-marker helpers its inline tests call, and the test module; each
`impl Checker for <Lang>Code` moves to src/checker/<lang>.rs.

Pure code movement: bodies verbatim, no logic or metric-output change.
The 11 local macros resolve into the child files via textual scoping
(mod decls after the macro_rules!), so no macro-export change and no
visibility widening were needed. Per-call macro rationale comments
(e.g. PHP's String3 hidden-supertype note on impl_simple_is_string!)
moved with their invocations into the language files, per
.claude/rules/macro-comments.md. lib.rs is unchanged (`mod checker;` +
`pub(crate) use crate::checker::*;` — the per-language impls are
unnamed and not re-exported; the shared pub(crate) helpers stay in the
parent so the glob and the direct crate::checker:: paths from
getter/{java,csharp}.rs keep resolving).

checker.rs 2038 -> 463 production SLOC; 25 per-language modules added.
