+++ src/checker.rs
@@ -318,391 +318,31 @@ pub(crate) trait Checker {
     }
 }
 
-impl Checker for PreprocCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Preproc::Comment
-    }
-
-    impl_simple_is_string!(Preproc, StringLiteral, RawStringLiteral);
-}
-
-impl Checker for CcommentCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Ccomment::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        get_aho_corasick_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    impl_simple_is_string!(Ccomment, StringLiteral, RawStringLiteral);
-}
-
-impl Checker for CppCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Cpp::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        get_aho_corasick_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    // Issue #285 contract: every `Cpp::FunctionDefinition*` alias must
-    // be enumerated here AND in `is_func`, `get_func_space_name`, and
-    // `get_space_kind` (see `src/getter.rs`). Aliased kind_ids
-    // 489/491/494 are not emitted by the currently pinned
-    // `tree-sitter-cpp` parse tables on any input we can construct,
-    // so a missing variant won't fail a parse-and-assert test — it
-    // will silently drop those nodes from FuncSpace creation the next
-    // time a grammar bump starts emitting them (see lesson 2 in
-    // `docs/development/lessons_learned.md`).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Cpp::TranslationUnit
-                | Cpp::FunctionDefinition
-                | Cpp::FunctionDefinition2
-                | Cpp::FunctionDefinition3
-                | Cpp::FunctionDefinition4
-                | Cpp::StructSpecifier
-                | Cpp::ClassSpecifier
-                | Cpp::NamespaceDefinition
-        )
-    }
-
-    // Issue #285 contract: keep this in sync with `is_func_space` and
-    // the C++ getters — see comment above `is_func_space`.
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Cpp::FunctionDefinition
-                | Cpp::FunctionDefinition2
-                | Cpp::FunctionDefinition3
-                | Cpp::FunctionDefinition4
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Cpp::LambdaExpression
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Cpp::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Cpp::LPAREN | Cpp::LPAREN2 | Cpp::COMMA | Cpp::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(Cpp, StringLiteral, ConcatenatedString, RawStringLiteral);
-
-    impl_is_else_if_parent_clause!(Cpp, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == Cpp::PrimitiveType
-    }
-}
-
-impl Checker for CCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == C::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        get_aho_corasick_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    // C has no classes/namespaces and no methods, so the only code
-    // spaces are the translation unit and function definitions. Struct /
-    // union / enum specifiers hold no functions in C, so — unlike the
-    // C++ checker — they are deliberately NOT spaces (including them
-    // would create empty FuncSpaces for every aggregate type).
-    // `FunctionDefinition2` is the grammar's `function_definition` alias
-    // (kind_id 197); enumerate it alongside the primary so neither is
-    // silently dropped from FuncSpace creation (#285, lesson 2).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            C::TranslationUnit | C::FunctionDefinition | C::FunctionDefinition2
-        )
-    }
-
-    // Keep in sync with `is_func_space` and the C getters (#285).
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            C::FunctionDefinition | C::FunctionDefinition2
-        )
-    }
-
-    // C has no closures/lambdas.
-    fn is_closure(_node: &Node) -> bool {
-        false
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == C::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(node.kind_id().into(), C::LPAREN | C::COMMA | C::RPAREN)
-    }
-
-    // C has no raw string literals.
-    impl_simple_is_string!(C, StringLiteral, ConcatenatedString);
-
-    impl_is_else_if_parent_clause!(C, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == C::PrimitiveType
-    }
-}
-
-impl Checker for ObjcCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Objc::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        get_aho_corasick_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    // Objective-C builds on C, so the code spaces are the translation
-    // unit, free functions, and ObjC-specific containers: `@interface`
-    // / `@protocol` (declarations only → `Interface`), `@implementation`
-    // (method bodies → `Class`), and method definitions. `BlockLiteral`
-    // (`^{ … }`) is a closure counted by `nom`, not its own space —
-    // mirroring how the C++ checker treats `LambdaExpression`. Keep this
-    // list in sync with `is_func`, `get_func_space_name`, and
-    // `get_space_kind` (see `src/getter.rs`); the `FunctionDefinition2`
-    // alias must ride alongside the primary or it is silently dropped
-    // from FuncSpace creation (#285, lesson 2).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Objc::TranslationUnit
-                | Objc::FunctionDefinition
-                | Objc::FunctionDefinition2
-                | Objc::MethodDefinition
-                | Objc::ClassInterface
-                | Objc::ClassImplementation
-                | Objc::ProtocolDeclaration
-        )
-    }
-
-    // Keep in sync with `is_func_space` and the ObjC getters (#285).
-    // A `method_definition` is the `@implementation`-side method with a
-    // body; the `@interface`-side `method_declaration` has no body and
-    // is therefore not a function.
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Objc::FunctionDefinition | Objc::FunctionDefinition2 | Objc::MethodDefinition
-        )
-    }
-
-    // ObjC blocks `^{ … }` are the language's closures.
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Objc::BlockLiteral
-    }
-
-    // A C `call_expression` (aliased to `CallExpression2`) or an ObjC
-    // message send `[receiver method]` are both calls.
-    fn is_call(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Objc::CallExpression | Objc::CallExpression2 | Objc::MessageExpression
-        )
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Objc::LPAREN | Objc::LPAREN2 | Objc::COMMA | Objc::RPAREN
-        )
-    }
-
-    // ObjC has no raw string literals; `@"…"` parses as `AT` + a plain
-    // `string_literal`, so the same string kinds as C apply.
-    impl_simple_is_string!(Objc, StringLiteral, ConcatenatedString);
-
-    impl_is_else_if_parent_clause!(Objc, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == Objc::PrimitiveType
-    }
-}
-
-impl Checker for MozcppCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Mozcpp::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        get_aho_corasick_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    // Issue #285 contract: every `Mozcpp::FunctionDefinition*` alias must
-    // be enumerated here AND in `is_func`, `get_func_space_name`, and
-    // `get_space_kind` (see `src/getter.rs`). Aliased kind_ids
-    // 489/491/494 are not emitted by the currently pinned
-    // `tree-sitter-mozcpp` parse tables on any input we can construct,
-    // so a missing variant won't fail a parse-and-assert test — it
-    // will silently drop those nodes from FuncSpace creation the next
-    // time a grammar bump starts emitting them (see lesson 2 in
-    // `docs/development/lessons_learned.md`).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Mozcpp::TranslationUnit
-                | Mozcpp::FunctionDefinition
-                | Mozcpp::FunctionDefinition2
-                | Mozcpp::FunctionDefinition3
-                | Mozcpp::FunctionDefinition4
-                | Mozcpp::StructSpecifier
-                | Mozcpp::ClassSpecifier
-                | Mozcpp::NamespaceDefinition
-        )
-    }
-
-    // Issue #285 contract: keep this in sync with `is_func_space` and
-    // the C++ getters — see comment above `is_func_space`.
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Mozcpp::FunctionDefinition
-                | Mozcpp::FunctionDefinition2
-                | Mozcpp::FunctionDefinition3
-                | Mozcpp::FunctionDefinition4
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Mozcpp::LambdaExpression
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Mozcpp::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Mozcpp::LPAREN | Mozcpp::LPAREN2 | Mozcpp::COMMA | Mozcpp::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(Mozcpp, StringLiteral, ConcatenatedString, RawStringLiteral);
-
-    impl_is_else_if_parent_clause!(Mozcpp, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == Mozcpp::PrimitiveType
-    }
-}
-
-impl Checker for PythonCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Python::Comment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        // comment containing coding info are useful
-        node.start_row() <= 1
-            && RE
-                .get_or_init(|| {
-                    Regex::new(r"^[ \t\f]*#.*?coding[:=][ \t]*([-_.a-zA-Z0-9]+)")
-                        .expect("constant Python coding-declaration regex always compiles")
-                })
-                .is_match(&code[node.start_byte()..node.end_byte()])
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Python::Module | Python::FunctionDefinition | Python::ClassDefinition
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        node.kind_id() == Python::FunctionDefinition
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        // Route through the single lambda-alias chokepoint so closure
-        // detection here and the cognitive lambda-scope walks accept the
-        // exact same set: `Lambda` (196, the concrete production emitted
-        // today) and `Lambda2` (197, the currently-unseen hidden alias).
-        // `Lambda3` (73, the `lambda` keyword token) is intentionally
-        // excluded. Centralizing means a future grammar bump that
-        // promotes `Lambda2` cannot silently undercount closures in
-        // nom/nargs or desync from cognitive (issues #419/#422; lesson 2
-        // in lessons_learned.md). The drift-guard test below asserts
-        // `Lambda2` stays unseen until then.
-        crate::metrics::cognitive::python_is_lambda(node)
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Python::Call
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        // tree-sitter-python emits the PEP 570 positional-only marker `/`
-        // as a `positional_separator` node and the PEP 3102 keyword-only
-        // marker `*` as a `keyword_separator` node, both as direct children
-        // of the `parameters` list. They are punctuation, not parameters, so
-        // they must be excluded or they inflate nargs by one each (issue #414).
-        matches!(
-            node.kind_id().into(),
-            Python::LPAREN
-                | Python::COMMA
-                | Python::RPAREN
-                | Python::PositionalSeparator
-                | Python::KeywordSeparator
-        )
-    }
-
-    impl_simple_is_string!(Python, String, ConcatenatedString);
-
-    // Python models `elif` as a dedicated `elif_clause` node, which is
-    // handled directly by cognitive/cyclomatic dispatch as a branch
-    // extension — `is_else_if` is intentionally never invoked for
-    // `elif_clause` because it is not an `if_statement` and is not in
-    // any of the structural kind sets that `count_specific_ancestors`
-    // walks for nesting (issue #274).
-    //
-    // `else: if x: ...` chains also exist — semantically equivalent to
-    // `else if` — but the grammar wraps the inner `if_statement` in a
-    // `block` node, so the shape is `else_clause → block → if_statement`
-    // rather than the direct `else_clause → if_statement` used by
-    // C++/JS/TS/TSX/Rust. Match the chained shape by walking through
-    // the `block` and requiring the inner `if` to be the block's sole
-    // named child; sibling statements would mean a real nested-if, not
-    // a chain (issue #276).
-    //
-    // `block` has two aliased kind_ids in tree-sitter-python
-    // (`Block` = 135, `Block2` = 160 — both surface as `"block"`); we
-    // accept either via `python_is_block`, the single normalization
-    // point for the aliases (issue #419; lesson 2 in
-    // `docs/development/lessons_learned.md`).
-    fn is_else_if(node: &Node) -> bool {
-        node.kind_id() == Python::IfStatement
-            && node.parent().is_some_and(|parent| {
-                crate::metrics::npa::python_is_block(&parent)
-                    && parent.children().filter(Node::is_named).count() == 1
-                    && parent
-                        .parent()
-                        .is_some_and(|gp| gp.kind_id() == Python::ElseClause)
-            })
-    }
-}
+mod bash;
+mod c;
+mod ccomment;
+mod cpp;
+mod csharp;
+mod elixir;
+mod go;
+mod groovy;
+mod irules;
+mod java;
+mod javascript;
+mod kotlin;
+mod lua;
+mod mozcpp;
+mod mozjs;
+mod objc;
+mod perl;
+mod php;
+mod preproc;
+mod python;
+mod ruby;
+mod rust;
+mod tcl;
+mod tsx;
+mod typescript;
 
 /// Returns the `class_body` child of a Java `object_creation_expression`
 /// when the node is an anonymous class (`new Runnable() { ... }`), or
@@ -713,74 +353,6 @@ impl Checker for PythonCode {
 /// reaches this path.
 pub(crate) fn java_anonymous_class_body<'a>(node: &Node<'a>) -> Option<Node<'a>> {
     node.first_child(|id| id == Java::ClassBody as u16)
-}
-
-impl Checker for JavaCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Java::LineComment || node.kind_id() == Java::BlockComment
-    }
-
-    // `EnumDeclaration`, `RecordDeclaration`, and `AnnotationTypeDeclaration`
-    // are class-like declarations that can contain fields and methods,
-    // so they open a class space alongside `ClassDeclaration` /
-    // `InterfaceDeclaration` (issue #280). Without them, `Npa`/`Npm`/`Wmc`
-    // never see their bodies as class scopes and silently produce zero
-    // counts. Annotation types map to `Interface` in `get_space_kind`
-    // (their elements are abstract methods at the bytecode level);
-    // enums and records map to `Class`.
-    //
-    // An `object_creation_expression` carrying a `class_body` child is
-    // an anonymous class (`new Runnable() { ... }`); it opens its own
-    // Class space so its members are attributed to it, not the
-    // enclosing method (#463). A plain `new Foo()` has no `class_body`
-    // child and must not open a space, so the arm is gated on the
-    // body's presence. This mirrors PHP's `AnonymousClass` handling and
-    // brings Java to parity with PHP/C# anonymous forms.
-    fn is_func_space(node: &Node) -> bool {
-        if node.kind_id() == Java::ObjectCreationExpression as u16 {
-            return java_anonymous_class_body(node).is_some();
-        }
-        matches!(
-            node.kind_id().into(),
-            Java::Program
-                | Java::ClassDeclaration
-                | Java::InterfaceDeclaration
-                | Java::EnumDeclaration
-                | Java::RecordDeclaration
-                | Java::AnnotationTypeDeclaration
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        node.kind_id() == Java::MethodDeclaration || node.kind_id() == Java::ConstructorDeclaration
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Java::LambdaExpression
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Java::MethodInvocation
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        // Java's explicit receiver parameter (`void m(S this, int a)`, JLS
-        // 8.4.1) parses as a `receiver_parameter` child of
-        // `formal_parameters`, distinct from a real `formal_parameter`. It
-        // binds `this`, not a value, so it is not a formal parameter and
-        // must be excluded — matching Rust's `SelfParameter` (#457), Go's
-        // `receiver` field, and C++'s implicit `this` (#470).
-        matches!(
-            node.kind_id().into(),
-            Java::LPAREN | Java::COMMA | Java::RPAREN | Java::ReceiverParameter
-        )
-    }
-
-    impl_simple_is_string!(Java, StringLiteral, MultilineStringLiteral);
-
-    // tree-sitter-java models `else if` as an `Else` keyword token followed
-    // by a nested `if_statement` (no wrapping `else_clause` node).
-    impl_is_else_if_prev_sibling!(Java, IfStatement, Else);
 }
 
 /// Counts the `accessor_declaration` children (`get` / `set` / `init`) of a C#
@@ -813,290 +385,6 @@ pub(crate) fn csharp_accessor_count(node: &Node) -> usize {
 /// accessor-less expression-bodied form (#464 indexer, #472 property).
 pub(crate) fn csharp_member_has_accessors(node: &Node) -> bool {
     csharp_accessor_count(node) > 0
-}
-
-impl Checker for CsharpCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Csharp::Comment
-    }
-
-    // A bodied indexer (`this[int i] { get; set; }`) or property
-    // (`int X { get; set; }`) defers to its `accessor_declaration` children
-    // for its function spaces — counting it here too double-counts (property,
-    // #472) or triple-counts (indexer, #464) the member in nom/wmc. Only the
-    // accessor-less expression-bodied form (`this[int i] => _d[i];` /
-    // `int W => _w;`) opens a space directly, matching the npm `.max(1)`
-    // fallback and the way the implicit getter is the sole callable.
-    fn is_func_space(node: &Node) -> bool {
-        if matches!(
-            node.kind_id().into(),
-            Csharp::IndexerDeclaration | Csharp::PropertyDeclaration
-        ) {
-            return !csharp_member_has_accessors(node);
-        }
-        matches!(
-            node.kind_id().into(),
-            Csharp::CompilationUnit
-                | Csharp::ClassDeclaration
-                | Csharp::StructDeclaration
-                | Csharp::RecordDeclaration
-                | Csharp::InterfaceDeclaration
-                | Csharp::EnumDeclaration
-                | Csharp::MethodDeclaration
-                | Csharp::ConstructorDeclaration
-                | Csharp::DestructorDeclaration
-                | Csharp::LocalFunctionStatement
-                | Csharp::LambdaExpression
-                | Csharp::AnonymousMethodExpression
-                | Csharp::AccessorDeclaration
-                | Csharp::OperatorDeclaration
-                | Csharp::ConversionOperatorDeclaration
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        if matches!(
-            node.kind_id().into(),
-            Csharp::IndexerDeclaration | Csharp::PropertyDeclaration
-        ) {
-            return !csharp_member_has_accessors(node);
-        }
-        matches!(
-            node.kind_id().into(),
-            Csharp::MethodDeclaration
-                | Csharp::ConstructorDeclaration
-                | Csharp::DestructorDeclaration
-                | Csharp::LocalFunctionStatement
-                | Csharp::AccessorDeclaration
-                | Csharp::OperatorDeclaration
-                | Csharp::ConversionOperatorDeclaration
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Csharp::LambdaExpression | Csharp::AnonymousMethodExpression
-        )
-    }
-
-    fn is_call(node: &Node) -> bool {
-        // The C# grammar emits three aliased `kind_id`s for
-        // `invocation_expression`; matching only the unsuffixed variant
-        // silently drops the rest (lesson #2 in lessons_learned.md).
-        matches!(node.kind_id().into(), csharp_invocation_expr_kinds!())
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Csharp::LPAREN | Csharp::COMMA | Csharp::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(
-        Csharp,
-        StringLiteral,
-        VerbatimStringLiteral,
-        RawStringLiteral,
-        InterpolatedStringExpression,
-    );
-
-    // tree-sitter-c-sharp models `else if` as an `Else` keyword token
-    // followed by a nested `if_statement` (no wrapping `else_clause` node).
-    impl_is_else_if_prev_sibling!(Csharp, IfStatement, Else);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        // Without this, every `PredefinedType` keyword (`int`, `string`,
-        // `bool`, `object`, …) collapses into a single Halstead operator
-        // because they share one `kind_id`. Returning `true` here routes
-        // them through the lexeme-keyed `primitive_operators` map so
-        // distinct keywords count as distinct operators (issue #286).
-        node.kind_id() == Csharp::PredefinedType as u16
-    }
-}
-
-impl Checker for MozjsCode {
-    fn is_comment(node: &Node) -> bool {
-        // `html_comment` is the ECMAScript Annex-B `<!-- -->` form, a
-        // distinct named+extra grammar node (Mozjs 133 / JS 132 / TS 162
-        // / Tsx 168) alongside the `//` and `/* */` `comment` kind. It
-        // must classify as a comment for loc/tokens/find parity (#697).
-        matches!(node.kind_id().into(), Mozjs::Comment | Mozjs::HtmlComment)
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Mozjs::Program
-                | Mozjs::FunctionExpression
-                | Mozjs::Class
-                | Mozjs::GeneratorFunction
-                | Mozjs::FunctionDeclaration
-                | Mozjs::MethodDefinition
-                | Mozjs::GeneratorFunctionDeclaration
-                | Mozjs::ClassDeclaration
-                | Mozjs::ArrowFunction
-        )
-    }
-
-    is_js_func_and_closure_checker!(Mozjs);
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Mozjs::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Mozjs::LPAREN | Mozjs::COMMA | Mozjs::RPAREN
-        )
-    }
-
-    impl_js_family_is_string!(Mozjs);
-
-    impl_is_else_if_parent_clause!(Mozjs, IfStatement, ElseClause);
-}
-
-impl Checker for JavascriptCode {
-    fn is_comment(node: &Node) -> bool {
-        // See MozjsCode::is_comment — `html_comment` (Annex-B `<!-- -->`)
-        // is a comment kind for loc/tokens/find parity (#697).
-        matches!(
-            node.kind_id().into(),
-            Javascript::Comment | Javascript::HtmlComment
-        )
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Javascript::Program
-                | Javascript::FunctionExpression
-                | Javascript::Class
-                | Javascript::GeneratorFunction
-                | Javascript::FunctionDeclaration
-                | Javascript::MethodDefinition
-                | Javascript::GeneratorFunctionDeclaration
-                | Javascript::ClassDeclaration
-                | Javascript::ArrowFunction
-        )
-    }
-
-    is_js_func_and_closure_checker!(Javascript);
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Javascript::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Javascript::LPAREN | Javascript::COMMA | Javascript::RPAREN
-        )
-    }
-
-    impl_js_family_is_string!(Javascript);
-
-    impl_is_else_if_parent_clause!(Javascript, IfStatement, ElseClause);
-}
-
-impl Checker for TypescriptCode {
-    fn is_comment(node: &Node) -> bool {
-        // See MozjsCode::is_comment — `html_comment` (Annex-B `<!-- -->`)
-        // is a comment kind for loc/tokens/find parity (#697).
-        matches!(
-            node.kind_id().into(),
-            Typescript::Comment | Typescript::HtmlComment
-        )
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Typescript::Program
-                | Typescript::FunctionExpression
-                | Typescript::Class
-                | Typescript::GeneratorFunction
-                | Typescript::FunctionDeclaration
-                | Typescript::MethodDefinition
-                | Typescript::GeneratorFunctionDeclaration
-                | Typescript::ClassDeclaration
-                | Typescript::AbstractClassDeclaration
-                | Typescript::InterfaceDeclaration
-                | Typescript::ArrowFunction
-        )
-    }
-
-    is_js_func_and_closure_checker!(Typescript);
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Typescript::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Typescript::LPAREN | Typescript::COMMA | Typescript::RPAREN
-        )
-    }
-
-    impl_js_family_is_string!(Typescript);
-
-    impl_is_else_if_parent_clause!(Typescript, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == Typescript::PredefinedType
-    }
-}
-
-impl Checker for TsxCode {
-    fn is_comment(node: &Node) -> bool {
-        // See MozjsCode::is_comment — `html_comment` (Annex-B `<!-- -->`)
-        // is a comment kind for loc/tokens/find parity (#697).
-        matches!(node.kind_id().into(), Tsx::Comment | Tsx::HtmlComment)
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Tsx::Program
-                | Tsx::FunctionExpression
-                | Tsx::Class
-                | Tsx::GeneratorFunction
-                | Tsx::FunctionDeclaration
-                | Tsx::MethodDefinition
-                | Tsx::GeneratorFunctionDeclaration
-                | Tsx::ClassDeclaration
-                | Tsx::AbstractClassDeclaration
-                | Tsx::InterfaceDeclaration
-                | Tsx::ArrowFunction
-        )
-    }
-
-    is_js_func_and_closure_checker!(Tsx);
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Tsx::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Tsx::LPAREN | Tsx::COMMA | Tsx::RPAREN
-        )
-    }
-
-    impl_js_family_is_string!(Tsx, String3);
-
-    impl_is_else_if_parent_clause!(Tsx, IfStatement, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        node.kind_id() == Tsx::PredefinedType
-    }
 }
 
 /// Strip the `#` / `#!` marker plus the `[...]` brackets from a
@@ -1160,869 +448,6 @@ fn rust_inner_attr_marks_test(node: &Node, code: &[u8]) -> bool {
         }
     }
     false
-}
-
-impl Checker for RustCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Rust::LineComment || node.kind_id() == Rust::BlockComment
-    }
-
-    fn is_useful_comment(node: &Node, code: &[u8]) -> bool {
-        if let Some(parent) = node.parent()
-            && parent.kind_id() == Rust::TokenTree
-        {
-            // A comment could be a macro token
-            return true;
-        }
-        let code = &code[node.start_byte()..node.end_byte()];
-        code.starts_with(b"/// cbindgen:")
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Rust::SourceFile
-                | Rust::FunctionItem
-                | Rust::ImplItem
-                | Rust::TraitItem
-                | Rust::ClosureExpression
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        node.kind_id() == Rust::FunctionItem
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Rust::ClosureExpression
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Rust::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        // A *typed* receiver (`self: Box<Self>`, `self: Rc<Self>`,
-        // `self: Pin<&mut Self>` — arbitrary self types) does not parse as
-        // `SelfParameter`; the grammar models it as an ordinary `parameter`
-        // node whose binding is the `self` keyword (`Rust::Zelf`). It is
-        // still a receiver, so it is excluded too, for parity with the
-        // bare-receiver case and with Go/C++ (#457). A normal `parameter`
-        // such as `x: i32` binds an `identifier`, never `self`, so this
-        // child check is unambiguous.
-        let is_typed_self_receiver = node.kind_id() == Rust::Parameter
-            && node.children().any(|child| child.kind_id() == Rust::Zelf);
-
-        // `SelfParameter` is Rust's bare method receiver (`self`, `&self`,
-        // `&mut self`). Like Go's `receiver` field and C++'s implicit
-        // `this`, it is not a formal parameter and must not be counted
-        // (see #457).
-        matches!(
-            node.kind_id().into(),
-            Rust::LPAREN
-                | Rust::COMMA
-                | Rust::RPAREN
-                | Rust::PIPE
-                | Rust::AttributeItem
-                | Rust::SelfParameter
-        ) || is_typed_self_receiver
-    }
-
-    impl_simple_is_string!(Rust, StringLiteral, RawStringLiteral);
-
-    // Rust models `else if` as a nested `if_expression` (not an
-    // `if_statement`) sitting directly inside the `else_clause`.
-    impl_is_else_if_parent_clause!(Rust, IfExpression, ElseClause);
-
-    #[inline]
-    fn is_primitive(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Rust::PrimitiveType
-                | Rust::PrimitiveType2
-                | Rust::PrimitiveType3
-                | Rust::PrimitiveType4
-                | Rust::PrimitiveType5
-                | Rust::PrimitiveType6
-                | Rust::PrimitiveType7
-                | Rust::PrimitiveType8
-                | Rust::PrimitiveType9
-                | Rust::PrimitiveType10
-                | Rust::PrimitiveType11
-                | Rust::PrimitiveType12
-                | Rust::PrimitiveType13
-                | Rust::PrimitiveType14
-                | Rust::PrimitiveType15
-                | Rust::PrimitiveType16
-                | Rust::PrimitiveType17
-        )
-    }
-
-    /// Skip the subtree when `node` is a `mod`, `fn`, `impl`,
-    /// `trait`, `const`, or `static` item marked test-only by an
-    /// outer or inner attribute (`#[test]`, `#[cfg(test)]`,
-    /// `#[tokio::test]`, `#![cfg(test)]`, …). The runtime guard
-    /// in `spaces::metrics_with_options` only consults this hook
-    /// when the caller opts in via `MetricsOptions::exclude_tests`,
-    /// so the default `metrics()` entry point is unaffected.
-    fn should_skip_subtree(node: &Node, code: &[u8]) -> bool {
-        if !matches!(
-            node.kind_id().into(),
-            Rust::ModItem
-                | Rust::FunctionItem
-                | Rust::ImplItem
-                | Rust::TraitItem
-                | Rust::ConstItem
-                | Rust::StaticItem
-        ) {
-            return false;
-        }
-        rust_item_is_test_only(node, code)
-    }
-}
-
-impl Checker for GoCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Go::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Go::SourceFile | Go::FunctionDeclaration | Go::MethodDeclaration | Go::FuncLiteral
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Go::FunctionDeclaration | Go::MethodDeclaration
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Go::FuncLiteral
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Go::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(node.kind_id().into(), Go::LPAREN | Go::COMMA | Go::RPAREN)
-    }
-
-    impl_simple_is_string!(Go, InterpretedStringLiteral, RawStringLiteral);
-
-    #[inline]
-    fn is_else_if(node: &Node) -> bool {
-        node.kind_id() == Go::IfStatement
-            && node
-                .parent()
-                .is_some_and(|p| p.kind_id() == Go::IfStatement)
-    }
-}
-
-impl Checker for KotlinCode {
-    fn is_comment(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Kotlin::LineComment | Kotlin::BlockComment
-        )
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Kotlin::SourceFile
-                | Kotlin::ClassDeclaration
-                | Kotlin::ObjectDeclaration
-                | Kotlin::CompanionObject
-                | Kotlin::ObjectLiteral
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Kotlin::FunctionDeclaration | Kotlin::SecondaryConstructor
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Kotlin::LambdaLiteral | Kotlin::AnonymousFunction
-        )
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Kotlin::CallExpression
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Kotlin::LPAREN | Kotlin::COMMA | Kotlin::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(Kotlin, StringLiteral, MultilineStringLiteral);
-
-    // tree-sitter-kotlin models `else if` as an `else` keyword sibling
-    // followed by an `if_expression` (not an `if_statement`, and no
-    // wrapping clause node).
-    impl_is_else_if_prev_sibling!(Kotlin, IfExpression, Else);
-}
-
-impl Checker for PerlCode {
-    fn is_comment(node: &Node) -> bool {
-        matches!(node.kind_id().into(), Perl::Comments | Perl::PodStatement)
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Perl::SourceFile
-                | Perl::FunctionDefinition
-                | Perl::FunctionDefinitionWithoutSub
-                | Perl::AnonymousFunction
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Perl::FunctionDefinition | Perl::FunctionDefinitionWithoutSub
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Perl::AnonymousFunction
-    }
-
-    fn is_call(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Perl::CallExpressionWithSpacedArgs
-                | Perl::CallExpressionWithSub
-                | Perl::CallExpressionWithArgsWithBrackets
-                | Perl::CallExpressionWithVariable
-                | Perl::CallExpressionWithBareword
-                | Perl::CallExpressionRecursive
-                | Perl::MethodInvocation
-        )
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Perl::LPAREN | Perl::COMMA | Perl::RPAREN | Perl::FatComma
-        )
-    }
-
-    // `HeredocBodyStatement` wraps the heredoc body text (and any
-    // `Interpolation` children) that appears as a top-level
-    // statement after the heredoc-introducing `<<TAG`; it is the
-    // visible literal node and is treated as a string here, the
-    // same way Bash's `heredoc_body` is treated as a string.
-    impl_simple_is_string!(
-        Perl,
-        StringSingleQuoted,
-        StringDoubleQuoted,
-        StringQQuoted,
-        StringQqQuoted,
-        BacktickQuoted,
-        CommandQxQuoted,
-        HeredocBodyStatement,
-    );
-
-    // tree-sitter-perl emits `elsif_clause` as a direct child of the
-    // surrounding `if_statement` (not as a wrapper around a nested
-    // `if`), so the clause node itself is the else-if.
-    impl_is_else_if_clause!(Perl, ElsifClause);
-}
-
-impl Checker for LuaCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Lua::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Lua::Chunk
-                | Lua::FunctionDeclaration
-                | Lua::FunctionDeclaration2
-                | Lua::FunctionDeclaration3
-                | Lua::FunctionDefinition
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Lua::FunctionDeclaration | Lua::FunctionDeclaration2 | Lua::FunctionDeclaration3
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Lua::FunctionDefinition
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Lua::FunctionCall
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        // NOTE: `impl NArgs for LuaCode` overrides `compute` with a positive
-        // filter on `Identifier | VarargExpression` and never calls `is_non_arg`.
-        // This implementation satisfies the trait contract but is unused for NArgs.
-        matches!(
-            node.kind_id().into(),
-            Lua::LPAREN | Lua::COMMA | Lua::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(Lua, String);
-
-    // Lua uses a dedicated elseif_statement node rather than nesting a
-    // second if_statement inside the outer one (as Go does).
-    impl_is_else_if_clause!(Lua, ElseifStatement);
-}
-
-impl Checker for BashCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Bash::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Bash::Program | Bash::FunctionDefinition
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        node.kind_id() == Bash::FunctionDefinition
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Bash::Command
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Bash::LPAREN | Bash::RPAREN | Bash::COMMA | Bash::SEMI
-        )
-    }
-
-    // tree-sitter-bash 0.25.1 only emits the `heredoc_body`
-    // parser-node symbol (`HeredocBody2`) in observed parse trees;
-    // the duplicate `HeredocBody` entry plus the hidden
-    // `_heredoc_body` (`HeredocBody3`) and `_simple_heredoc_body`
-    // (`SimpleHeredocBody`) rules do not surface, so they are
-    // intentionally omitted here.
-    impl_simple_is_string!(
-        Bash,
-        String,
-        RawString,
-        AnsiCString,
-        TranslatedString,
-        HeredocBody2,
-    );
-
-    // tree-sitter-bash exposes a dedicated `elif_clause` node, so the
-    // clause node itself is the else-if (no nested `if` to inspect).
-    impl_is_else_if_clause!(Bash, ElifClause);
-}
-
-impl Checker for TclCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Tcl::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(node.kind_id().into(), Tcl::SourceFile | Tcl::Procedure)
-    }
-
-    fn is_func(node: &Node) -> bool {
-        node.kind_id() == Tcl::Procedure
-    }
-
-    // Tcl closures (`apply`) are ordinary commands; the grammar has no distinct closure node.
-    fn is_closure(_: &Node) -> bool {
-        false
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Tcl::Command
-    }
-
-    // Tcl arguments are whitespace-separated; no punctuation to exclude.
-    fn is_non_arg(_: &Node) -> bool {
-        false
-    }
-
-    impl_simple_is_string!(Tcl, QuotedWord, BracedWord, BracedWordSimple);
-
-    // Tcl grammar has a dedicated `elseif` named node, not a nested `if`.
-    impl_is_else_if_clause!(Tcl, Elseif);
-}
-
-impl Checker for IrulesCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Irules::Comment
-    }
-
-    // iRules code is dominated by `when EVENT { … }` event handlers; real
-    // scripts frequently contain no `proc` at all. Treating `when` handlers
-    // and `proc` definitions as function spaces is what makes per-function
-    // metrics meaningful here — under a proc-only model a typical
-    // handler-only file would collapse to a single Unit space. The handler
-    // bodies are recursively-parsed `braced_word`s, so nested control flow
-    // is attributed correctly.
-    //
-    // `OnHandler` / `TrapHandler` are defensive arms (lesson #34): the
-    // grammar declares dedicated `on_handler` / `trap_handler` nodes, but at
-    // statement level a bare `on …` / `trap …` parses as a generic `command`
-    // (the `command` rule wins), so these kinds do not surface in practice
-    // and have no dedicated test. They are listed here so that if a future
-    // grammar revision starts emitting them they are handled as spaces
-    // rather than silently dropped — never removed in favour of an
-    // `unreachable!` (lesson #5).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Irules::SourceFile
-                | Irules::Procedure
-                | Irules::WhenEvent
-                | Irules::OnHandler
-                | Irules::TrapHandler
-        )
-    }
-
-    // Handlers count as functions (not closures): `nom.functions` on a
-    // typical iRules file is then the handler count, the intuitive metric.
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Irules::Procedure | Irules::WhenEvent | Irules::OnHandler | Irules::TrapHandler
-        )
-    }
-
-    // iRules has no anonymous lambda node (`apply` is an ordinary command).
-    fn is_closure(_: &Node) -> bool {
-        false
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Irules::Command
-    }
-
-    // iRules arguments are whitespace-separated; no punctuation to exclude.
-    fn is_non_arg(_: &Node) -> bool {
-        false
-    }
-
-    impl_simple_is_string!(Irules, QuotedWord, BracedWord, BracedWordSimple);
-
-    // iRules grammar has a dedicated `elseif` named node (id 145), not a
-    // nested `if`. `Elseif2` (id 92) is the `elseif` keyword token, not
-    // the clause, so it is intentionally excluded here (lesson #34
-    // defensive-arm marker).
-    impl_is_else_if_clause!(Irules, Elseif);
-}
-
-impl Checker for PhpCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Php::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Php::Program
-                | Php::FunctionDefinition
-                | Php::MethodDeclaration
-                | Php::AnonymousFunction
-                | Php::ArrowFunction
-                | Php::ClassDeclaration
-                | Php::InterfaceDeclaration
-                | Php::TraitDeclaration
-                | Php::EnumDeclaration
-                | Php::AnonymousClass
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Php::FunctionDefinition | Php::MethodDeclaration
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Php::AnonymousFunction | Php::ArrowFunction
-        )
-    }
-
-    // Intentionally narrower than ABC's `branches` set: ABC additionally
-    // counts `ObjectCreationExpression` (`new Foo()`) as a branch, but
-    // `is_call` drives the `--ops` CLI feature and should match the
-    // user's mental model of "function/method call sites" (mirrors
-    // Java's `is_call` = `MethodInvocation` while ABC counts
-    // `MethodInvocation | New`).
-    fn is_call(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Php::FunctionCallExpression
-                | Php::MemberCallExpression
-                | Php::ScopedCallExpression
-                | Php::NullsafeMemberCallExpression
-        )
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Php::LPAREN | Php::LPAREN2 | Php::COMMA | Php::RPAREN | Php::RPAREN2 | Php::DOTDOTDOT
-        )
-    }
-
-    // `String` is the named single-quoted literal; `String2` and
-    // `String3` are aliased kind_ids that the language enum also
-    // maps to `"string"` (`String2` is the `string` type keyword
-    // and `String3` is the hidden `_string` supertype that covers
-    // any string literal). Include all three so generic
-    // string-filtering stays consistent with `get_op_type` and the
-    // `Alterator` text-preservation arm (issue #288).
-    impl_simple_is_string!(
-        Php,
-        String,
-        String2,
-        String3,
-        EncapsedString,
-        Heredoc,
-        Nowdoc,
-        ShellCommandExpression,
-    );
-
-    // PHP models `else if` in two distinct shapes, so it cannot use a
-    // single is_else_if macro:
-    //
-    // 1. The one-word `elseif` keyword parses as a dedicated
-    //    `else_if_clause` node (`ElseIfClause`); both the brace
-    //    `} elseif {` and the alternative `elseif: … endif;` colon forms
-    //    parse as `ElseIfClause`. `ElseIfClause2` is an aliased kind_id
-    //    the grammar maps to the same `else_if_clause` rule name but which
-    //    does not surface in observed parse trees; it is kept as a
-    //    defensive arm (lesson #34) so a future grammar revision that emits
-    //    it is handled rather than silently dropped.
-    // 2. The two-word `else if` form (only valid in the brace syntax)
-    //    parses as an `else_clause` wrapping a nested `if_statement`
-    //    (`else_clause → if_statement`, the `impl_is_else_if_parent_clause!`
-    //    shape used by C++/JS/Rust). The inner `IfStatement` must be
-    //    recognized as an else-if continuation, or the cognitive
-    //    `IfStatement` arm double-counts it and inflates nesting for later
-    //    arms (#529). A plain `else { if (…) {} }` does NOT match: there the
-    //    inner `if` nests under a `compound_statement`, not directly under
-    //    the `else_clause`.
-    //
-    // The alternative colon syntax (`if …: … endif;`) only accepts the
-    // one-word `elseif` (shape 1); two-word `else if` there is a PHP fatal
-    // parse error and the grammar emits an `ERROR` node, so it is
-    // intentionally not handled here — there is no well-defined metric for
-    // syntactically invalid source.
-    //
-    // Matching shapes 1 and 2 lets the cognitive guard
-    // `IfStatement if !Self::is_else_if(node)` suppress the nested-if
-    // penalty while the wrapping `else_clause` still scores its +1
-    // branch extension. The dedicated-clause match (shape 1) is
-    // behaviorally inert today — `count_specific_ancestors` is not used by
-    // PHP cognitive and the guard only fires on `IfStatement` — but is kept
-    // for parity with the PHP cyclomatic/ABC/cognitive dispatch which list
-    // both clause variants.
-    #[inline]
-    fn is_else_if(node: &Node) -> bool {
-        let kind = node.kind_id();
-        matches!(kind.into(), Php::ElseIfClause | Php::ElseIfClause2)
-            || (kind == Php::IfStatement
-                && node.parent().is_some_and(|parent| {
-                    matches!(parent.kind_id().into(), Php::ElseClause | Php::ElseClause2)
-                }))
-    }
-}
-
-impl Checker for ElixirCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Elixir::Comment
-    }
-
-    // Elixir has no syntactic function-definition node: `def`/`defp` /
-    // `defmacro`/`defmacrop` / `defmodule` are ordinary `Call` nodes
-    // with the macro identifier in the `target` field. The byte-less
-    // [`is_func_space`] and [`is_func`] cannot distinguish them from
-    // any other `Call`, so they conservatively return zero (only the
-    // file root and explicit anonymous functions surface as func
-    // spaces). The text-aware [`is_func_space_with_code`] /
-    // [`is_func_with_code`] overrides below promote the macro-shaped
-    // declarations to first-class function / class spaces (#275). The
-    // walker passes the source bytes through, so the metrics
-    // attributed to a `def`'s body now correctly nest under a Function
-    // space and `Wmc` / `Npm` / `Npa` see a `defmodule` Class.
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Elixir::Source | Elixir::AnonymousFunction
-        )
-    }
-
-    fn is_func(_: &Node) -> bool {
-        false
-    }
-
-    fn is_func_space_with_code(node: &Node, code: &[u8]) -> bool {
-        use crate::metrics::cognitive::{
-            elixir_call_keyword, elixir_is_class_macro, elixir_is_inside_quote_block,
-            elixir_is_method_macro,
-        };
-        if Self::is_func_space(node) {
-            return true;
-        }
-        let Some(kw) = elixir_call_keyword(node, code) else {
-            return false;
-        };
-        if elixir_is_class_macro(kw) {
-            return true;
-        }
-        // A `def` / `defp` / `defmacro` / `defmacrop` nested inside a
-        // `quote do … end` template does NOT declare a method of any
-        // enclosing module — the syntax tree there is a code template
-        // emitted later, on macro expansion (#310).
-        elixir_is_method_macro(kw) && !elixir_is_inside_quote_block(node, code)
-    }
-
-    fn is_func_with_code(node: &Node, code: &[u8]) -> bool {
-        use crate::metrics::cognitive::{
-            elixir_call_keyword, elixir_is_inside_quote_block, elixir_is_method_macro,
-        };
-        let Some(kw) = elixir_call_keyword(node, code) else {
-            return false;
-        };
-        elixir_is_method_macro(kw) && !elixir_is_inside_quote_block(node, code)
-    }
-
-    fn promotes_to_func_space_with_code(node: &Node, code: &[u8]) -> bool {
-        use crate::metrics::cognitive::{
-            elixir_call_keyword, elixir_is_class_macro, elixir_is_inside_quote_block,
-            elixir_is_method_macro,
-        };
-        // Cheap path: byte-less `is_func_space` matches `Source` and
-        // `AnonymousFunction` without needing any text inspection.
-        if Self::is_func_space(node) {
-            return true;
-        }
-        // Otherwise one `elixir_call_keyword` lookup answers the
-        // combined question instead of the two the default impl would
-        // have made via `is_func_with_code || is_func_space_with_code`.
-        let Some(kw) = elixir_call_keyword(node, code) else {
-            return false;
-        };
-        if elixir_is_class_macro(kw) {
-            return true;
-        }
-        elixir_is_method_macro(kw) && !elixir_is_inside_quote_block(node, code)
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        node.kind_id() == Elixir::AnonymousFunction
-    }
-
-    fn is_call(node: &Node) -> bool {
-        node.kind_id() == Elixir::Call
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Elixir::LPAREN | Elixir::LPAREN2 | Elixir::RPAREN | Elixir::COMMA
-        )
-    }
-
-    impl_simple_is_string!(Elixir, String, Charlist, Sigil);
-
-    // Elixir lacks an `else if` chain construct. Multi-way branching uses
-    // `cond do ... end` (a `Call` whose `do_block` holds many
-    // `stab_clause`s, each `+1` nesting in cognitive) or nested
-    // `if/else`. No tail-recursive chain to collapse here.
-    #[inline]
-    fn is_else_if(_: &Node) -> bool {
-        false
-    }
-}
-
-impl Checker for RubyCode {
-    fn is_comment(node: &Node) -> bool {
-        node.kind_id() == Ruby::Comment
-    }
-
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Ruby::Program
-                | Ruby::Method
-                | Ruby::SingletonMethod
-                | Ruby::Lambda
-                | Ruby::Block
-                | Ruby::DoBlock
-                | Ruby::Class
-                | Ruby::SingletonClass
-                | Ruby::Module
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(node.kind_id().into(), Ruby::Method | Ruby::SingletonMethod)
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        match node.kind_id().into() {
-            Ruby::Lambda => true,
-            // A stabby lambda `->(z) { … }` parses as a `Lambda` node that
-            // CONTAINS the `Block`/`DoBlock` for its body, so the `Lambda`
-            // arm above already counts it. Counting the inner block again
-            // would double-count one closure as two (#465). The keyword
-            // forms `lambda { }` / `proc { }` parse as a `Call` carrying a
-            // `Block`/`DoBlock` argument (parent is not a `Lambda`), so they
-            // still count exactly once.
-            Ruby::Block | Ruby::DoBlock => node
-                .parent()
-                .is_none_or(|parent| parent.kind_id() != Ruby::Lambda),
-            _ => false,
-        }
-    }
-
-    // tree-sitter-ruby 0.23.1 emits four aliased visible variants of the
-    // `call` rule (`Call`, `Call2`, `Call3`, `Call4`); `Call5` ("_call")
-    // is the hidden inner production and does not surface.
-    fn is_call(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Ruby::Call | Ruby::Call2 | Ruby::Call3 | Ruby::Call4
-        )
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        // `PIPE` is included because block parameter lists are delimited
-        // by `|` rather than parentheses (e.g. `[1,2,3].each { |x| … }`).
-        matches!(
-            node.kind_id().into(),
-            Ruby::LPAREN
-                | Ruby::LPAREN2
-                | Ruby::RPAREN
-                | Ruby::RPAREN2
-                | Ruby::COMMA
-                | Ruby::SEMI
-                | Ruby::PIPE
-        )
-    }
-
-    impl_simple_is_string!(
-        Ruby,
-        String,
-        ChainedString,
-        BareString,
-        Subshell,
-        Regex,
-        HeredocBody,
-        DelimitedSymbol,
-        SimpleSymbol,
-        StringArray,
-        SymbolArray,
-        Character,
-    );
-
-    // tree-sitter-ruby exposes `elsif` as its own named clause node, so the
-    // dedicated-clause-node strategy applies here (same as Lua/Bash/PHP).
-    impl_is_else_if_clause!(Ruby, Elsif);
-}
-
-impl Checker for GroovyCode {
-    fn is_comment(node: &Node) -> bool {
-        // `groovydoc_comment` is Groovy's `/** … */` Groovydoc kind, a
-        // distinct grammar node (Groovy 150) from the `//` LineComment
-        // and `/* */` BlockComment. The `Loc` cloc arm already counts it
-        // (loc.rs); this restores parity so tokens/find agree too (#697).
-        matches!(
-            node.kind_id().into(),
-            Groovy::LineComment | Groovy::BlockComment | Groovy::GroovydocComment
-        )
-    }
-
-    // Mirrors `impl Checker for JavaCode` exactly: `EnumDeclaration`,
-    // `RecordDeclaration`, and `AnnotationTypeDeclaration` open class
-    // spaces so `Npa`/`Npm`/`Wmc` walk their bodies. Cross-language parity
-    // with Java is the point — identical-shape sources must agree on
-    // class-shaped metrics (issue #280).
-    fn is_func_space(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Groovy::SourceFile
-                | Groovy::ClassDeclaration
-                | Groovy::TraitDeclaration
-                | Groovy::InterfaceDeclaration
-                | Groovy::EnumDeclaration
-                | Groovy::RecordDeclaration
-                | Groovy::AnnotationTypeDeclaration
-        )
-    }
-
-    fn is_func(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Groovy::MethodDeclaration | Groovy::ConstructorDeclaration
-        )
-    }
-
-    fn is_closure(node: &Node) -> bool {
-        matches!(node.kind_id().into(), Groovy::Closure)
-    }
-
-    // `command_chain` is the new grammar's distinct node for Groovy's
-    // command-style juxtaposed calls (`foo bar baz`) which the prior
-    // amaanq grammar mis-modelled as `juxt_function_call`; it is a
-    // genuine method-call form and stays in `is_call`.
-    //
-    // Intentionally excludes `ObjectCreationExpression` (`new Foo()`):
-    // `is_call` follows the Java-family convention (Java's `is_call` =
-    // `MethodInvocation`, C#'s = `InvocationExpression`) of counting
-    // method/function call sites only. Constructor invocations are an
-    // ABC concern — Groovy's ABC `branches` counts `New` separately
-    // (see `groovy_count_token_branch` in metrics/abc.rs).
-    fn is_call(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Groovy::MethodInvocation | Groovy::CommandChain
-        )
-    }
-
-    fn is_non_arg(node: &Node) -> bool {
-        matches!(
-            node.kind_id().into(),
-            Groovy::LPAREN | Groovy::COMMA | Groovy::RPAREN
-        )
-    }
-
-    impl_simple_is_string!(Groovy, StringLiteral);
-
-    // The dekobon Groovy grammar models `if_statement` with the `else`
-    // keyword token emitted inline followed by the inner `if_statement`
-    // sibling — same shape as the prior amaanq grammar and Java.
-    impl_is_else_if_prev_sibling!(Groovy, IfStatement, Else);
 }
 
 #[cfg(test)]
