+++ src/checker/javascript.rs
@@ -1,0 +1,47 @@
+//! `Checker` implementation for JavaScript.
+#![allow(clippy::wildcard_imports, clippy::enum_glob_use)]
+
+use super::*;
+
+impl Checker for JavascriptCode {
+    fn is_comment(node: &Node) -> bool {
+        // See MozjsCode::is_comment — `html_comment` (Annex-B `<!-- -->`)
+        // is a comment kind for loc/tokens/find parity (#697).
+        matches!(
+            node.kind_id().into(),
+            Javascript::Comment | Javascript::HtmlComment
+        )
+    }
+
+    fn is_func_space(node: &Node) -> bool {
+        matches!(
+            node.kind_id().into(),
+            Javascript::Program
+                | Javascript::FunctionExpression
+                | Javascript::Class
+                | Javascript::GeneratorFunction
+                | Javascript::FunctionDeclaration
+                | Javascript::MethodDefinition
+                | Javascript::GeneratorFunctionDeclaration
+                | Javascript::ClassDeclaration
+                | Javascript::ArrowFunction
+        )
+    }
+
+    is_js_func_and_closure_checker!(Javascript);
+
+    fn is_call(node: &Node) -> bool {
+        node.kind_id() == Javascript::CallExpression
+    }
+
+    fn is_non_arg(node: &Node) -> bool {
+        matches!(
+            node.kind_id().into(),
+            Javascript::LPAREN | Javascript::COMMA | Javascript::RPAREN
+        )
+    }
+
+    impl_js_family_is_string!(Javascript);
+
+    impl_is_else_if_parent_clause!(Javascript, IfStatement, ElseClause);
+}
